Fix RangeError in messageDeleteBulk logging
If a message contained an attachment, embed, or anything of that nature, the bot failed to post the messageDeleteBulk log due to a RangeError. This was because the bot is not permitted to send empty content in an embed (by Discord), and attachments/embeds don't get pushed to fields, so empty content was the result. Anyway, this is the fix I wrote for it, if there's something cleaner instead, let me know.